# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry farm management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FlockPurpose = (typeof FlockPurpose)[keyof typeof FlockPurpose];

export const FlockPurpose = {
  eggs: "eggs",
  meat: "meat",
  hatching: "hatching",
  mixed: "mixed",
} as const;

export interface Flock {
  id: number;
  name: string;
  breed: string;
  count: number;
  ageWeeks: number;
  purpose: FlockPurpose;
  notes?: string | null;
  createdAt: string;
}

export type CreateFlockBodyPurpose =
  (typeof CreateFlockBodyPurpose)[keyof typeof CreateFlockBodyPurpose];

export const CreateFlockBodyPurpose = {
  eggs: "eggs",
  meat: "meat",
  hatching: "hatching",
  mixed: "mixed",
} as const;

export interface CreateFlockBody {
  name: string;
  breed: string;
  count: number;
  ageWeeks: number;
  purpose: CreateFlockBodyPurpose;
  notes?: string | null;
}

export type HatchingCycleStatus =
  (typeof HatchingCycleStatus)[keyof typeof HatchingCycleStatus];

export const HatchingCycleStatus = {
  incubating: "incubating",
  hatching: "hatching",
  completed: "completed",
  failed: "failed",
} as const;

export interface HatchingCycle {
  id: number;
  batchName: string;
  eggsSet: number;
  eggsHatched?: number | null;
  startDate: string;
  expectedHatchDate: string;
  actualHatchDate?: string | null;
  status: HatchingCycleStatus;
  temperature?: number | null;
  humidity?: number | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateHatchingCycleBodyStatus =
  (typeof CreateHatchingCycleBodyStatus)[keyof typeof CreateHatchingCycleBodyStatus];

export const CreateHatchingCycleBodyStatus = {
  incubating: "incubating",
  hatching: "hatching",
  completed: "completed",
  failed: "failed",
} as const;

export interface CreateHatchingCycleBody {
  batchName: string;
  eggsSet: number;
  eggsHatched?: number | null;
  startDate: string;
  expectedHatchDate: string;
  actualHatchDate?: string | null;
  status: CreateHatchingCycleBodyStatus;
  temperature?: number | null;
  humidity?: number | null;
  notes?: string | null;
}

export type TaskCategory = (typeof TaskCategory)[keyof typeof TaskCategory];

export const TaskCategory = {
  feeding: "feeding",
  health: "health",
  cleaning: "cleaning",
  hatching: "hatching",
  other: "other",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  category: TaskCategory;
  priority: TaskPriority;
  completed: boolean;
  dueDate?: string | null;
  createdAt: string;
}

export type CreateTaskBodyCategory =
  (typeof CreateTaskBodyCategory)[keyof typeof CreateTaskBodyCategory];

export const CreateTaskBodyCategory = {
  feeding: "feeding",
  health: "health",
  cleaning: "cleaning",
  hatching: "hatching",
  other: "other",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskBody {
  title: string;
  description?: string | null;
  category: CreateTaskBodyCategory;
  priority: CreateTaskBodyPriority;
  dueDate?: string | null;
}

export type UpdateTaskBodyCategory =
  (typeof UpdateTaskBodyCategory)[keyof typeof UpdateTaskBodyCategory];

export const UpdateTaskBodyCategory = {
  feeding: "feeding",
  health: "health",
  cleaning: "cleaning",
  hatching: "hatching",
  other: "other",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  category?: UpdateTaskBodyCategory;
  priority?: UpdateTaskBodyPriority;
  completed?: boolean;
  dueDate?: string | null;
}

export type GoalCategory = (typeof GoalCategory)[keyof typeof GoalCategory];

export const GoalCategory = {
  production: "production",
  health: "health",
  growth: "growth",
  financial: "financial",
  other: "other",
} as const;

export interface Goal {
  id: number;
  title: string;
  description?: string | null;
  targetValue: number;
  currentValue: number;
  unit: string;
  category: GoalCategory;
  deadline?: string | null;
  completed: boolean;
  createdAt: string;
}

export type CreateGoalBodyCategory =
  (typeof CreateGoalBodyCategory)[keyof typeof CreateGoalBodyCategory];

export const CreateGoalBodyCategory = {
  production: "production",
  health: "health",
  growth: "growth",
  financial: "financial",
  other: "other",
} as const;

export interface CreateGoalBody {
  title: string;
  description?: string | null;
  targetValue: number;
  currentValue: number;
  unit: string;
  category: CreateGoalBodyCategory;
  deadline?: string | null;
}

export type UpdateGoalBodyCategory =
  (typeof UpdateGoalBodyCategory)[keyof typeof UpdateGoalBodyCategory];

export const UpdateGoalBodyCategory = {
  production: "production",
  health: "health",
  growth: "growth",
  financial: "financial",
  other: "other",
} as const;

export interface UpdateGoalBody {
  title?: string;
  description?: string | null;
  targetValue?: number;
  currentValue?: number;
  unit?: string;
  category?: UpdateGoalBodyCategory;
  deadline?: string | null;
  completed?: boolean;
}

export type ActivityLogCategory =
  (typeof ActivityLogCategory)[keyof typeof ActivityLogCategory];

export const ActivityLogCategory = {
  feeding: "feeding",
  health: "health",
  hatching: "hatching",
  cleaning: "cleaning",
  observation: "observation",
  other: "other",
} as const;

export interface ActivityLog {
  id: number;
  title: string;
  description?: string | null;
  category: ActivityLogCategory;
  date: string;
  createdAt: string;
}

export type CreateActivityLogBodyCategory =
  (typeof CreateActivityLogBodyCategory)[keyof typeof CreateActivityLogBodyCategory];

export const CreateActivityLogBodyCategory = {
  feeding: "feeding",
  health: "health",
  hatching: "hatching",
  cleaning: "cleaning",
  observation: "observation",
  other: "other",
} as const;

export interface CreateActivityLogBody {
  title: string;
  description?: string | null;
  category: CreateActivityLogBodyCategory;
  date: string;
}

export interface DashboardSummary {
  totalFlocks: number;
  totalChickens: number;
  activeHatchingCycles: number;
  totalEggsIncubating: number;
  tasksDueToday: number;
  tasksCompletedToday: number;
  goalsCompleted: number;
  totalGoals: number;
  overallHatchRate: number;
}

export interface HatchStats {
  totalCycles: number;
  completedCycles: number;
  totalEggsSet: number;
  totalEggsHatched: number;
  averageHatchRate: number;
  bestHatchRate: number;
  recentCycles: HatchingCycle[];
}

export type ListTasksParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
};