# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry farm management API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityLogCategory } from "./activityLogCategory";

export interface ActivityLog {
  id: number;
  title: string;
  description?: string | null;
  category: ActivityLogCategory;
  date: string;
  createdAt: string;
}