# lib/api-zod/src/generated/types/createGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry farm management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGoalBodyCategory } from "./createGoalBodyCategory";

export interface CreateGoalBody {
  title: string;
  description?: string | null;
  targetValue: number;
  currentValue: number;
  unit: string;
  category: CreateGoalBodyCategory;
  deadline?: string | null;
}