# lib/api-zod/src/generated/types/createHatchingCycleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry farm management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateHatchingCycleBodyStatus } from "./createHatchingCycleBodyStatus";

export interface CreateHatchingCycleBody {
  batchName: string;
  eggsSet: number;
  eggsHatched?: number | null;
  startDate: string;
  expectedHatchDate: string;
  actualHatchDate?: string | null;
  status: CreateHatchingCycleBodyStatus;
  temperature?: number | null;
  humidity?: number | null;
  notes?: string | null;
}