# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry farm management API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalFlocks: number;
  totalChickens: number;
  activeHatchingCycles: number;
  totalEggsIncubating: number;
  tasksDueToday: number;
  tasksCompletedToday: number;
  goalsCompleted: number;
  totalGoals: number;
  overallHatchRate: number;
}